backend/easyocr-service/requirements.txt
flask==2.3.3
easyocr==1.7.0
pillow==10.0.1
numpy==1.24.3
opencv-python==4.8.1.78
torch==2.0.1
torchvision==0.15.2
